--- server/src/common/dto/auth/login-user.dto.ts
@@ -1,7 +1,0 @@
-import { PickType } from '@nestjs/swagger';
-import { CreateLocalUserRequest } from '@common/requests/user';
-
-export class LoginRequest extends PickType(CreateLocalUserRequest, [
-  'email',
-  'password',
-] as const) {}
