--- server/src/common/dto/ranking/pagination-request.dto.ts
@@ -1,9 +1,0 @@
-import { IsNumber } from 'class-validator';
-
-export class PaginationRequestDto {
-  @IsNumber()
-  page: number;
-
-  @IsNumber()
-  limit: number;
-}
