--- server/src/common/dto/ranking/pagination-response.dto.ts
@@ -1,4 +1,0 @@
-export class PaginationResponseDto {
-  totalPage: number;
-  nextPage: number;
-}
