--- server/src/common/dto/ranking/user-info-with-rank.dto.ts
@@ -1,15 +1,0 @@
-import { ApiProperty, PickType } from '@nestjs/swagger';
-import { UserProfileDto } from '../user/user-profile.dto';
-
-export class UserInfoWithRankDto extends PickType(UserProfileDto, [
-  'id',
-  'nickname',
-  'point',
-] as const) {
-  @ApiProperty({
-    example: 1,
-    description: '순위',
-    required: true,
-  })
-  rank: number;
-}
