+++ server/src/common/dto/user-info/index.ts
@@ -1,0 +1,1 @@
+export * from './user-info-with-rank.dto';
