+++ server/src/common/dto/user-info/user-info-with-rank.dto.ts
@@ -1,0 +1,31 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+export class UserInfoWithRankDto {
+  @ApiProperty({
+    example: 1,
+    description: 'id',
+    required: true,
+  })
+  id: number;
+
+  @ApiProperty({
+    example: 'test',
+    description: '닉네임',
+    required: true,
+  })
+  nickname: string;
+
+  @ApiProperty({
+    example: 0,
+    description: '포인트 점수',
+    required: true,
+  })
+  point: number;
+
+  @ApiProperty({
+    example: 1,
+    description: '순위',
+    required: true,
+  })
+  rank: number;
+}
